# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Harency ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/get_card": {
            "get": {
                "tags": [
                    "Card"
                ],
                "summary": "Get card details",
                "description": "Get card details",
                "operationId": "authGetCard",
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong invalid credentials. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/delete_card": {
            "post": {
                "tags": [
                    "Card"
                ],
                "summary": "Delete Card",
                "description": "Delete user card",
                "operationId": "authDeleteCard",
                "requestBody": {
                    "description": "Pass user card details Id",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "card_id"
                                ],
                                "properties": {
                                    "card_id": {
                                        "type": "string",
                                        "format": "card_id",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong card details. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/add_card": {
            "post": {
                "tags": [
                    "Card"
                ],
                "summary": "Add Card",
                "description": "Create user card",
                "operationId": "authCreateCard",
                "requestBody": {
                    "description": "Pass user card details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "card_holder_name",
                                    "card_number",
                                    "expiry_date",
                                    "cvv",
                                    "card_type",
                                    "default_card"
                                ],
                                "properties": {
                                    "card_holder_name": {
                                        "type": "string",
                                        "format": "card_holder_name",
                                        "example": "john walker"
                                    },
                                    "card_number": {
                                        "type": "string",
                                        "format": "card_number",
                                        "example": "4242424242424242"
                                    },
                                    "expiry_date": {
                                        "type": "string",
                                        "format": "expiry_date",
                                        "example": "08/26"
                                    },
                                    "cvv": {
                                        "type": "string",
                                        "format": "cvv",
                                        "example": "111"
                                    },
                                    "card_type": {
                                        "type": "string",
                                        "format": "card_type",
                                        "example": "visa"
                                    },
                                    "default_card": {
                                        "type": "string",
                                        "format": "default_card",
                                        "example": "primary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong card details. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}